Setup client with react and created NavBar

+++ server/controllers/product.controller.js
@@ -37,12 +37,18 @@ export const getProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product ID" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product Deleted" });
   } catch (error) {
     console.log("error in deleting product: ", error.message);
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
@@ -53,7 +59,7 @@ export const updateProduct = async (req, res) => {
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res
-      .status(400)
+      .status(404)
       .json({ success: false, message: "Invalid Product ID" });
   }
